some important fixes
